Reverted back to using activities for logging in and added logOut feature
