+++ app/src/main/java/edu/rosehulman/improse/myapplication/MainActivity.kt
@@ -15,6 +15,7 @@ import com.google.android.material.navigation.NavigationView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import android.view.Menu
+import android.view.MenuItem
 import android.view.View
 import android.widget.Toast
 import com.google.firebase.auth.FirebaseAuth
@@ -23,18 +24,9 @@ import com.google.firebase.firestore.ListenerRegistration
 import edu.rosehulman.rosefire.Rosefire
 
 
-class MainActivity : AppCompatActivity(),  LoginFragment.OnLoginPressed{
+class MainActivity : AppCompatActivity(){
 
     private lateinit var appBarConfiguration: AppBarConfiguration
-    private lateinit var listenerRegistration: ListenerRegistration
-    var showAll: Boolean = false;
-    private val RC_SIGN_IN = 1
-    private val ROSE_SIGN_IN = 2
-    var userId:String? = null
-
-
-    private val auth = FirebaseAuth.getInstance()
-    lateinit var authStateListener:FirebaseAuth.AuthStateListener
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -52,47 +44,25 @@ class MainActivity : AppCompatActivity(),  LoginFragment.OnLoginPressed{
         setupActionBarWithNavController(navController, appBarConfiguration)
         navView.setupWithNavController(navController)
 
-        if(savedInstanceState == null) {
-            initializeAuthListener()
-        }
     }
 
-    fun initializeAuthListener() {
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        // Handle action bar item clicks here. The action bar will
+        // automatically handle clicks on the Home/Up button, so long
+        // as you specify a parent activity in AndroidManifest.xml.
+        return when (item.itemId) {
 
-        Log.d("user", "initialize listener")
 
-        authStateListener = FirebaseAuth.AuthStateListener { auth ->
-            Log.d("user", "made listener")
-            val user = auth.currentUser
-            if(user != null){
-                userId = user.uid
-                Log.d("user", "user not null")
-               // switchToHomeFragment(user.uid)
+            R.id.action_settings -> {
+                return true
             }
-            else{
-                Log.d("user", "need to login")
-                switchToLoginFragment()
+            R.id.logout_setting ->{
+                val intent = Intent(this, LoginActivity::class.java)
+                startActivity(intent)
+                return true
             }
+            else -> super.onOptionsItemSelected(item)
         }
-
-    }
-
-    private fun switchToLoginFragment() {
-        val ft = supportFragmentManager.beginTransaction()
-        ft.replace(R.id.content_layout, LoginFragment.newInstance())
-        //ft.addToBackStack("lif")
-        ft.commit()
-    }
-
-    override fun onStart() {
-        super.onStart()
-        auth.addAuthStateListener(authStateListener)
-
-    }
-
-    override fun onStop() {
-        super.onStop()
-        auth.removeAuthStateListener(authStateListener)
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
@@ -105,44 +75,4 @@ class MainActivity : AppCompatActivity(),  LoginFragment.OnLoginPressed{
         return navController.navigateUp(appBarConfiguration) || super.onSupportNavigateUp()
     }
 
-
-    override fun roseLogin() {
-        onRosefireLogin()
-    }
-
-    fun onRosefireLogin() {
-        val signInIntent = Rosefire.getSignInIntent(this, getString(R.string.reg_token))
-        startActivityForResult(signInIntent, ROSE_SIGN_IN)
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (requestCode == ROSE_SIGN_IN) {
-            val result = Rosefire.getSignInResultFromIntent(data)
-            if (!result.isSuccessful) { // The user cancelled the login
-                return
-            }
-            FirebaseAuth.getInstance().signInWithCustomToken(result.token)
-                .addOnCompleteListener(
-                    this
-                ) { task ->
-                    Log.d("rose_fire",
-                        "signInWithCustomToken:onComplete:" + task.isSuccessful
-                    )
-                    // If sign in fails, display a message to the user. If sign in succeeds
-                    // you should use an AuthStateListener to handle the logic for
-                    // signed in user and a signed out user.
-                    if (!task.isSuccessful) {
-                        Log.w("rose_fire",
-                            "signInWithCustomToken",
-                            task.exception
-                        )
-                        Toast.makeText(
-                            this@MainActivity, "Authentication failed.",
-                            Toast.LENGTH_SHORT
-                        ).show()
-                    }
-                }
-        }
-    }
 }
