+++ app/src/main/res/values/strings.xml
@@ -58,8 +58,7 @@
     </string>
     <string name="rose_login_below">Use the button below to login using your Rose-Hulman credentials</string>
     <string name="reg_token">1b46204c-449e-43b9-984f-b019b474082d</string>
-
-
+    <string name="logout">Log Out</string>
 
 
 </resources>
